Moved test cases
